requirements.txt
alembic==1.3.2
Click==7.0
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.10.4
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.12
SQLAlchemy-Utils==0.36.1
Werkzeug==0.16.0
flask_cors==4.0.0
requests==2.31.0
openai
pinecone-client==3.2.2
bs4==0.0.2
